Set to 0/40 when the schedule is not active

+++ plugwise/helper.py
@@ -1130,8 +1130,10 @@ class SmileHelper:
                     entry = directive.find("then").attrib
                     keys, dummy = zip(*entry.items())
                     if str(keys[0]) == "preset":
-                        if loc_id == NONE:  # set to 0 when the schedule is not active
-                            schedule[directive.attrib["time"]] = [float(0), float(0)]
+                        if (
+                            loc_id == NONE
+                        ):  # set to 0/40 when the schedule is not active
+                            schedule[directive.attrib["time"]] = [float(0), float(40)]
                         else:
                             schedule[directive.attrib["time"]] = [
                                 float(self._presets(loc_id)[entry["preset"]][0]),
